Prefer reload to restart for HAProxy
The DFaaS agent restarts the HAProxy by issuing `systemctl restart
haproxy.service`, this causes a hard stop of the HAProxy, closing all
open connections and losing requests. The solution is to issue
`systemctl reload haproxy.service`, this will instruct HAProxy to do a
graceful stop when restarting. The latter will not close any open
connections and the restart mechanism without loss of requests is
automatically done by HAProxy in master-worker mode (the default on
Ubuntu).